Added a new npm task 'build-docker'
The new task 'build-docker' combines 'npm run build' with
'docker build -f Dockerfile.nginx -t single-page-app:latest .'
and thus provides a ready to use nginx Docker image with which to host
the built and now static single-page-application.
Not yet done is to execute the 'npm run build' within a Docker container,
currently the user has to have the node-packages installed on the host system.